requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
amqp==5.0.9
aniso8601==9.0.1
anyjson==0.3.3
apturl==0.5.2
astroid==2.11.7
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
bcrypt==3.2.0
billiard==3.6.4.0
black==23.1.0
blinker==1.4
boto3==1.26.47
botocore==1.29.47
Brlapi==0.8.3
cachelib==0.10.0
celery==5.2.3
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.1.1
click==8.0.3
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
command-not-found==0.3
contourpy==1.0.6
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
dill==0.3.5.1
distro==1.7.0
distro-info===1.1build1
dnspython==2.2.1
duplicity==0.8.21
email-validator==1.3.0
entrypoints==0.4
ephem==4.1.3
exceptiongroup==1.1.1
executing==1.1.1
fasteners==0.14.1
Flask==2.2.2
Flask-Caching==2.0.1
Flask-Login==0.6.2
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security-Too==5.0.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.2
github==1.2.7
graphviz==0.14.2
greenlet==2.0.1
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.0.0
ipykernel==6.16.1
ipython==8.5.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jupyter_client==7.4.3
jupyter_core==4.11.2
keyring==23.5.0
kiwisolver==1.4.4
kombu==5.2.3
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.7.1
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
mando==0.6.4
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.0
numpy==1.23.5
oauthlib==3.2.0
olefile==0.46
packaging==23.0
pandas==1.5.2
paramiko==2.9.3
parso==0.8.3
passlib==1.7.4
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipdeptree==2.2.0
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.31
protobuf==3.12.4
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycups==2.0.1
Pygments==2.13.0
PyGObject==3.42.1
PyJWT==2.3.0
pylint==2.14.4
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==7.2.2
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-memcached==1.58
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==24.0.1
radon==5.1.0
redis==4.4.1
reportlab==3.6.8
requests==2.25.1
s3transfer==0.6.0
scipy==1.10.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==1.4.44
stack-data==0.5.1
systemd-python==234
tomli==2.0.1
tomlkit==0.11.1
tornado==6.2
traitlets==5.5.0
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vine==5.0.0
wadllib==1.3.6
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
xdg==5
xkit==0.0.0
yarl==1.8.2
zipp==1.0.0
